Remove py26 support from castellan
As of mitaka, the infra team won't have the resources available to
reasonably test py26, also the oslo team is dropping py26 support
from their libraries. sine we rely on oslo for a lot of our work,
and depend on infra for our CI, we should drop py26 support too.

Change-Id: I4db3fb61038377bc96e47ac160a2da1652901a1b
Related-Bug: #1519510