Add bindep.txt
The Vault functional tests are currently failing in the gate, and it
appears to be because the unzip binary is not present. We need to
add it to bindep so it will be installed before running tests.

Change-Id: I4b8075429218a494f75f5e3611f48927d18f1700
Closes-Bug: 1840061
(cherry picked from commit 2547c8c9f8fa76b01aa133c53bc321bc2625668b)