vault: Add timeout option
The existing logic assumes that requests may raise timeout exception
but timeout exception is never raised unless timeout is explicitly set
in each request method calls. This introduces the timeout option and
use it in request method calls so that timeout is actually set and
handled by the logics.

Also the keystoneauth session options loaded in the vault driver is not
actually used. This change also removes it.

Change-Id: I8c354937c2c39a63f50c150b3e858826b6a78fe0