Make log levels more adequate
1) if some config option is not present in mapping file
it's totally ok. It should not be there. Log it only for debug reasons.
2) Castellan as config source is mostly used for sensitive data,
e.g. secrets, passwords, etc.
If some config option is specified in mapping file,
but Castellan id is not specified or is not present in Castellan
backend, then it's an error. Service will take default values and
produce awkward log messages about connection issues
to database, RPC service, Keystone.
There are no error mesasges from Castellan.

Closes-Bug: #2099924
Change-Id: Ib9be08cf6cda7d6f2c5ed7d014b40ceee11104af