Cleanup intermediate orders and containers
When creating keys or keypairs with Barbican, castellan creates
intermediate entities like orders and containers, yet returns to the
API caller only IDs of created secrets.

As a result, the API consumer is not aware of those entities and the
fact that they have to be deleted as well when secrets are deleted.
This leads to orders and containers piling up, and also the deleted
secrets can not be purged from DB as they are still being referenced
by not deleted containers and orders. This is especially pronounced
when the API caller is actually some other OpenStack service, so the
user may not even be aware that Barbican was involved in the first
place.

This patch attempts to delete those intermediate entities before
returning the secret IDs to the caller. For the sake of backward
compatibility, and to account for a fact that with default Barbican
policies, not every user who can create order can delete it too,
any errors on such attempts are ignored and only logged as warnings.

Closes-Bug: 1638018
Change-Id: Icdaa44a29fdc59fce107078bff3a7544beb2b3c7
Signed-off-by: Dejan Sanader <dejan.sanader@ovhcloud.com>