Fix hardcoded context string in _config_driver.py
The Castellan oslo.config driver is currently passing a literal string as the context to key manager get() method. This causes secret retrival to fail when using backends like Barbican, which needs a valid context.

This patch creates a valid context object from the configuration passed to the driver. The context is then used when the get() method is called, ensuring secrets can be fetched correctly.

Closes-Bug: #2126436
Change-Id: I6c5100abd927009de1f6b31f96096ce1296c86c3
Signed-off-by: abi-p <abi.perinparasa@canonical.com>